requirements.txt
langchain==0.1.20
langchain-core==0.1.52
langchain-community==0.0.38
langchain-text-splitters
langchain-openai
langchain_huggingface
langchain_chroma
langchain-groq
python-dotenv
ipykernel
pypdf
bs4
arxiv
pymupdf
wikipedia
chromadb
sentence_transformers
faiss-cpu
duckdb
pandas
openai
duckduckgo_search==5.3.1b1
pymupdf
arxiv
wikipedia
mysql-connector-python
SQLAlchemy
validators==0.28.1
unstructured
pytube
numexpr
huggingface_hub
youtube-transcript-api==0.6.1
pytube
streamlit>=1.30.0